requirements.txt
uvicorn==0.23.2
fastapi==0.105.0
python-multipart==0.0.6
pydantic_settings==2.1.0
python-jose==3.3.0
passlib==1.7.4